refactor and add paths
