+++ src/app/layout.tsx
@@ -1,3 +1,5 @@
+import Header from "@components/header";
+
 import type { Metadata } from "next";
 
 import localFont from "next/font/local";
@@ -21,7 +23,10 @@ export default function RootLayout({
 }) {
   return (
     <html lang="ar" dir="rtl">
-      <body className={sfArabicRounded.className}>{children}</body>
+      <body className={sfArabicRounded.className}>
+        <Header />
+        {children}
+      </body>
     </html>
   );
 }
