+++ src/components/writings/parts/WritingTile.tsx
@@ -1,4 +1,4 @@
-export default function WritingTile({ thing }: { thing: Thing }) {
+const WritingTile = ({ thing }: { thing: Thing }) => {
   return (
     <div className="h-96 rounded-2xl bg-orange-700 text-center">
       <img
@@ -9,4 +9,6 @@ export default function WritingTile({ thing }: { thing: Thing }) {
       <h3 className="text-base">{thing.title}</h3>
     </div>
   );
-}
+};
+
+export default WritingTile;
